requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
appdirs==1.4.4
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect2==0.7.0
discord.py==2.4.0
frozenlist==1.5.0
h11==0.14.0
idna==3.6
importlib_metadata==8.7.0
lxml==4.9.3
multidict==6.1.0
numpy==2.2.1
outcome==1.3.0.post0
pandas==2.2.3
Pillow==10.1.0
propcache==0.2.1
pycparser==2.21
pyee==11.1.1
PyPDF2==3.0.1
pyppeteer==2.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.0.8
requests==2.31.0
selenium==4.16.0
six==1.17.0
sniffio==1.3.0
sortedcontainers==2.4.0
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.67.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.13.2
tzdata==2024.2
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
